Refactoring - consolidated login close/open code in separate reusable functions
