+++ blocks/login/login.js
@@ -2,6 +2,35 @@ import {
   decorateIcons, makeLinksRelative, decorateButtons, getRootPath, decorateAnchor,
 } from '../../scripts/scripts.js';
 
+export function closeLoginModal() {
+  const loginEle = document.querySelector('.login-overlay');
+  const mainEle = document.querySelector('body');
+
+  loginEle.classList.remove('fade-in');
+  loginEle.classList.add('fade-out');
+
+  /* Delaying the removal of modal class to allow the fade-out animation to complete */
+  setTimeout(() => {
+    loginEle.classList.remove('modal');
+    mainEle.classList.remove('overflow-hidden');
+  }, 150);
+}
+
+export function openLoginModal() {
+  const loginEle = document.querySelector('.login-overlay');
+  const bodyEle = document.querySelector('body');
+
+  loginEle.classList.add('modal');
+  /* Delaying the fade-in class addition to allow the modal to be displayed */
+  setTimeout(() => {
+    loginEle.classList.remove('fade-out');
+    loginEle.classList.add('fade-in');
+  }, 150);
+  window.scrollTo(0, 0); // Scrolling to Top
+  bodyEle.classList.add('overflow-hidden');
+  bodyEle.classList.remove('overflowY-hidden');
+}
+
 export default async function decorate(block) {
   const resp = await fetch(`${getRootPath()}/login.plain.html`);
   if (resp && resp.status === 200) {
@@ -19,17 +48,7 @@ export default async function decorate(block) {
 
     if (overLayClose) {
       overLayClose.addEventListener('click', () => {
-        const loginEle = document.querySelector('.login-overlay');
-        const mainEle = document.querySelector('body');
-
-        loginEle.classList.remove('fade-in');
-        loginEle.classList.add('fade-out');
-
-        /* Delaying the removal of modal class to allow the fade-out animation to complete */
-        setTimeout(() => {
-          loginEle.classList.remove('modal');
-          mainEle.classList.remove('overflow-hidden');
-        }, 150);
+        closeLoginModal();
       });
     }
